2 test suite plus separate file for server
