+++ backend/src/index.ts
@@ -1,39 +1,11 @@
-import express, {Application} from "express";
-import bodyParser from "body-parser";
-import userRouter from "./users/views";
-import eventRouter from "./events/views";
-import swaggerUI from "swagger-ui-express";
-import spec from "../api-spec.json";
+import app from './server';
 
-const app: Application = express();
+const server = app.listen(process.env.PORT || 8000);
 
-// Middleware to parse json request bodies
-app.use(bodyParser.json());
-app.use("/api-docs", swaggerUI.serve, swaggerUI.setup(spec));
-
-/**
- * Sub-routers for our main router, we should have one sub-router per "entity" in the application
- */
-app.use("/user", userRouter);
-app.use("/events", eventRouter)
-
-// Root Url
-app.get("/", (req, res) => {
-  res.send("Hello World!").status(200);
-});
-
-// Default route for endpoints not defined
-app.get("*", (req, res) => {
-  res.send("You have reached a rounte not defined in this API");
-});
-
-app.post("/", (req, res) => {
-  res.send(req.body);
-});
-
-app.listen(process.env.PORT || 8000, async () => {
+server.on("listening", () => {
   console.log("✅ Server is up and running at http://localhost:8000");
 });
 
-export default app
-
+server.on("error", (error) => {
+  console.log("❌ Server failed to start due to error: %s", error);
+});
