+++ backend/src/server.ts
@@ -1,0 +1,34 @@
+import express, {Application} from "express";
+import bodyParser from "body-parser";
+import userRouter from "./users/views";
+import eventRouter from "./events/views";
+import swaggerUI from "swagger-ui-express";
+import spec from "../api-spec.json";
+
+const app: Application = express();
+
+// Middleware to parse json request bodies
+app.use(bodyParser.json());
+app.use("/api-docs", swaggerUI.serve, swaggerUI.setup(spec));
+
+/**
+ * Sub-routers for our main router, we should have one sub-router per "entity" in the application
+ */
+app.use("/user", userRouter);
+app.use("/events", eventRouter)
+
+// Root Url
+app.get("/", (req, res) => {
+  res.send("Hello World!").status(200);
+});
+
+// Default route for endpoints not defined
+app.get("*", (req, res) => {
+  res.send("You have reached a rounte not defined in this API");
+});
+
+app.post("/", (req, res) => {
+  res.send(req.body);
+});
+
+export default app
