+++ backend/src/users/controllers.ts
@@ -1,4 +1,3 @@
-import { Prisma } from "@prisma/client";
 import { Request, Response } from "express";
 
 // We are using one connection to prisma client to prevent multiple connections
